Implemented error handling for Shader program linking; More consistent usage of m_ naming convention; Moved Sphere transformation from Vertex shader to Fragment shader

+++ res/Model.glsl
@@ -16,7 +16,7 @@ const int GlassType = 1;
 
 struct Material
 {
-	float Type;
+	int Type;
 	vec3 Albedo;
 	float Roughness;
 	float Emission;
